Pull out the render function, so I can use it in other places.
